Extract CategorizationQueries collaborator
Moves the eight read-only reporting queries (get_active_categories,
get_all_categories, list_rules, list_merchants,
list_uncategorized_transactions, count_uncategorized,
categorization_stats, stats) and the CategorizationStats dataclass into a
new categorization/queries.py. These are presentation-shaped reads
consumed by the CLI and MCP surface — distinct from matcher.py's reads,
which serve the matching loop and return matcher-internal shapes.

CategorizationService.__init__ now constructs self._queries =
CategorizationQueries(db) and exposes the eight methods as thin
delegates. Facade drops from 1167 to 944 lines; the remaining bulk is the
orchestrator methods (categorize_items, _categorize_items_inner,
apply_rules, apply_merchant_categories, categorize_pending) — the next
extraction.

No behavior change; full suite passes.
