+++ src/moneybin/services/categorization/__init__.py
@@ -93,15 +93,25 @@ from moneybin.services.categorization.matcher import (
     CategorizationMatcher,
     match_merchants,
 )
+from moneybin.services.categorization.queries import (
+    CategorizationQueries,
+    CategorizationStats,
+)
 from moneybin.tables import (
     CATEGORIES,
-    CATEGORIZATION_RULES,
     FCT_TRANSACTIONS,
-    MERCHANTS,
-    TRANSACTION_CATEGORIES,
+)
+from moneybin.tables import (
+    CATEGORIZATION_RULES as CATEGORIZATION_RULES,
 )
 from moneybin.tables import (
     CATEGORY_OVERRIDES as CATEGORY_OVERRIDES,
+)
+from moneybin.tables import (
+    MERCHANTS as MERCHANTS,
+)
+from moneybin.tables import (
+    TRANSACTION_CATEGORIES as TRANSACTION_CATEGORIES,
 )
 from moneybin.tables import (
     USER_CATEGORIES as USER_CATEGORIES,
@@ -111,34 +121,6 @@ from moneybin.tables import (
 )
 
 logger = logging.getLogger(__name__)
-
-
-@dataclass(slots=True)
-class CategorizationStats:
-    """Typed result for categorization statistics."""
-
-    total: int
-    categorized: int
-    uncategorized: int
-    percent_categorized: float
-    by_source: dict[str, int]
-
-    def to_envelope(self) -> ResponseEnvelope:
-        """Build a ResponseEnvelope from this categorization stats result."""
-        data: dict[str, Any] = {
-            "total_transactions": self.total,
-            "categorized": self.categorized,
-            "uncategorized": self.uncategorized,
-            "percent_categorized": self.percent_categorized,
-            "by_source": self.by_source,
-        }
-        return build_envelope(
-            data=data,
-            sensitivity="low",
-            actions=[
-                "Use transactions_categorize_pending_list to see uncategorized transactions"
-            ],
-        )
 
 
 @dataclass(slots=True)
@@ -198,6 +180,7 @@ class CategorizationService:
         self._matcher = CategorizationMatcher(db)
         self._applier = MatchApplier(db, audit=self._audit)
         self._assist = AssistBridge(db)
+        self._queries = CategorizationQueries(db)
 
     # -- Per-transaction category writes (audit emission via applier) --
 
@@ -913,245 +896,39 @@ class CategorizationService:
 
     def get_active_categories(self) -> list[dict[str, str | bool | None]]:
         """Get all active categories."""
-        try:
-            rows = self._db.execute(
-                f"""
-                SELECT category_id, category, subcategory, description,
-                       is_default, plaid_detailed
-                FROM {CATEGORIES.full_name}
-                WHERE is_active = true
-                ORDER BY category, subcategory
-                """
-            ).fetchall()
-        except duckdb.CatalogException:
-            return []
-
-        return [
-            {
-                "category_id": r[0],
-                "category": r[1],
-                "subcategory": r[2],
-                "description": r[3],
-                "is_default": r[4],
-                "plaid_detailed": r[5],
-            }
-            for r in rows
-        ]
+        return self._queries.get_active_categories()
 
     def get_all_categories(
         self, *, include_inactive: bool
     ) -> list[dict[str, str | bool | None]]:
-        """Get categories with consistent field shape including is_active.
-
-        Active-only views can use ``get_active_categories()`` to omit
-        ``is_active`` from each row; this method always includes it so the
-        MCP tool surface is consumer-friendly when toggling the include flag.
-        """
-        where = "" if include_inactive else "WHERE is_active = true"
-        try:
-            rows = self._db.execute(
-                f"""
-                SELECT category_id, category, subcategory, description,
-                       is_default, is_active, plaid_detailed
-                FROM {CATEGORIES.full_name}
-                {where}
-                ORDER BY category, subcategory
-                """  # noqa: S608  # constant clause, not user input
-            ).fetchall()
-        except duckdb.CatalogException:
-            return []
-
-        return [
-            {
-                "category_id": r[0],
-                "category": r[1],
-                "subcategory": r[2],
-                "description": r[3],
-                "is_default": r[4],
-                "is_active": r[5],
-                "plaid_detailed": r[6],
-            }
-            for r in rows
-        ]
+        """Get categories with consistent field shape including is_active."""
+        return self._queries.get_all_categories(include_inactive=include_inactive)
 
     def list_rules(self) -> list[dict[str, Any]]:
         """List all categorization rules (active and inactive) ordered by priority."""
-        try:
-            rows = self._db.execute(
-                f"""
-                SELECT rule_id, name, merchant_pattern, match_type,
-                       min_amount, max_amount, account_id,
-                       category, subcategory, priority, is_active
-                FROM {CATEGORIZATION_RULES.full_name}
-                ORDER BY priority ASC, created_at ASC
-                """
-            ).fetchall()
-        except duckdb.CatalogException:
-            return []
-
-        return [
-            {
-                "rule_id": r[0],
-                "name": r[1],
-                "merchant_pattern": r[2],
-                "match_type": r[3],
-                "min_amount": r[4],
-                "max_amount": r[5],
-                "account_id": r[6],
-                "category": r[7],
-                "subcategory": r[8],
-                "priority": r[9],
-                "is_active": r[10],
-            }
-            for r in rows
-        ]
+        return self._queries.list_rules()
 
     def list_merchants(self) -> list[dict[str, str | None]]:
         """List all merchant name mappings ordered by canonical name."""
-        try:
-            rows = self._db.execute(
-                f"""
-                SELECT merchant_id, raw_pattern, match_type,
-                       canonical_name, category, subcategory
-                FROM {MERCHANTS.full_name}
-                ORDER BY canonical_name
-                """
-            ).fetchall()
-        except duckdb.CatalogException:
-            return []
-
-        return [
-            {
-                "merchant_id": r[0],
-                "raw_pattern": r[1],
-                "match_type": r[2],
-                "canonical_name": r[3],
-                "category": r[4],
-                "subcategory": r[5],
-            }
-            for r in rows
-        ]
+        return self._queries.list_merchants()
 
     def list_uncategorized_transactions(
         self, *, limit: int
     ) -> list[dict[str, Any]] | None:
-        """List uncategorized transactions ordered by date descending.
-
-        Returns ``None`` (rather than ``[]``) when the underlying tables don't
-        exist yet — callers can distinguish "no transactions" from "no schema"
-        and surface a more useful action hint.
-        """
-        try:
-            result = self._db.execute(
-                f"""
-                SELECT t.transaction_id, t.transaction_date, t.amount,
-                       t.description, t.memo, t.account_id
-                FROM {FCT_TRANSACTIONS.full_name} t
-                LEFT JOIN {TRANSACTION_CATEGORIES.full_name} c
-                    ON t.transaction_id = c.transaction_id
-                WHERE c.transaction_id IS NULL
-                ORDER BY t.transaction_date DESC
-                LIMIT ?
-                """,
-                [limit],
-            )
-            columns = [desc[0] for desc in result.description]
-            rows = result.fetchall()
-        except duckdb.CatalogException:
-            return None
-
-        return [dict(zip(columns, row, strict=False)) for row in rows]
+        """List uncategorized transactions ordered by date descending."""
+        return self._queries.list_uncategorized_transactions(limit=limit)
 
     def count_uncategorized(self) -> int:
         """Return the number of transactions without a category assignment."""
-        try:
-            row = self._db.execute(
-                f"""
-                SELECT COUNT(*) FROM {FCT_TRANSACTIONS.full_name} t
-                LEFT JOIN {TRANSACTION_CATEGORIES.full_name} c
-                    ON t.transaction_id = c.transaction_id
-                WHERE c.transaction_id IS NULL
-                """  # noqa: S608  # TableRef constants, no user input interpolated
-            ).fetchone()
-            return int(row[0]) if row else 0
-        except Exception:  # noqa: BLE001 — tables may not exist before first import
-            return 0
-
-    # -- Stats --
+        return self._queries.count_uncategorized()
 
     def categorization_stats(self) -> dict[str, int | float]:
-        """Get summary statistics about categorization coverage.
-
-        Returns:
-            Dict with total, categorized, uncategorized counts and
-            breakdown by categorized_by source.
-        """
-        try:
-            total_result = self._db.execute(
-                f"SELECT COUNT(*) FROM {FCT_TRANSACTIONS.full_name}"
-            ).fetchone()
-            total = total_result[0] if total_result else 0
-        except duckdb.CatalogException:
-            return {
-                "total": 0,
-                "categorized": 0,
-                "uncategorized": 0,
-                "pct_categorized": 0,
-            }
-
-        try:
-            categorized_result = self._db.execute(
-                f"SELECT COUNT(*) FROM {TRANSACTION_CATEGORIES.full_name}"
-            ).fetchone()
-            categorized = categorized_result[0] if categorized_result else 0
-        except duckdb.CatalogException:
-            categorized = 0
-
-        uncategorized = total - categorized
-        pct = round((categorized / total * 100), 1) if total > 0 else 0.0
-
-        stats: dict[str, int | float] = {
-            "total": total,
-            "categorized": categorized,
-            "uncategorized": uncategorized,
-            "pct_categorized": pct,
-        }
-
-        # Breakdown by source
-        try:
-            source_rows = self._db.execute(
-                f"""
-                SELECT categorized_by, COUNT(*) AS cnt
-                FROM {TRANSACTION_CATEGORIES.full_name}
-                GROUP BY categorized_by
-                ORDER BY cnt DESC
-                """
-            ).fetchall()
-            for source, count in source_rows:
-                stats[f"by_{source}"] = count
-        except duckdb.CatalogException:
-            pass
-
-        return stats
+        """Get summary statistics about categorization coverage."""
+        return self._queries.categorization_stats()
 
     def stats(self) -> CategorizationStats:
-        """Get categorization stats as a typed result.
-
-        Wrapper around :meth:`categorization_stats` that returns a typed object.
-        """
-        raw = self.categorization_stats()
-        by_source = {
-            k.removeprefix("by_"): v
-            for k, v in raw.items()
-            if k.startswith("by_") and isinstance(v, int)
-        }
-        return CategorizationStats(
-            total=int(raw["total"]),
-            categorized=int(raw["categorized"]),
-            uncategorized=int(raw["uncategorized"]),
-            percent_categorized=float(raw["pct_categorized"]),
-            by_source=by_source,
-        )
+        """Get categorization stats as a typed result."""
+        return self._queries.stats()
 
     def categorize_assist(
         self,
